move config back
